+++ CHANGELOG.md
@@ -2,6 +2,12 @@
 
 ## Unreleased
 
+### Changed
+
+  - Deployment modal - On update credential can be changed
+  - Deployment modal - On update only selected infra is visible
+  - Tools - Re-frame-10x is disabled by default
+
 ## [2.12.0] - 2020-12-10
 
 ### Changed